requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
plotly
pandas
python-dotenv
bcrypt
pydantic